Change flow_subjects controller to flows
